Add a test for checking the calculated cpu percent.
This change highlighted an issue with the tests that leaves intervals
running indefinitely for the duration of the test run.
This causes unusual behaviour, such as assertion blocks being called
multiple times.
To mitigate this during testing, a 'stop' feature was added to allow
stopping the cpu check interval for a specific component, thereby
allowing more control in tests